requirement.txt
langchain-openai>=0.0.6
langchain-core>=0.1.50
python-dotenv>=1.0.1
streamlit>=1.34.0
langchain-community>=0.0.34
langserve>=0.0.43
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.6.5
beautifulsoup4>=4.12.3
pypdf>=4.2.0
PyPDF2>=3.0.1
chromadb>=0.4.24
faiss-cpu>=1.7.4
groq>=0.4.0
cassio>=0.1.3
langchain-groq>=0.0.3
wikipedia>=1.4.0
arxiv>=1.4.7
langchainhub>=0.1.15
sentence-transformers>=2.5.1
langchain-objectbox>=0.0.6
